third_party/WebKit/LayoutTests/inspector/sources/debugger/file-system-project-mapping-expected.txt

Tests file system project mappings.


Running: testAutomaticMapping
Adding file system.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/bar.js.
Adding network resource.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/bar.js.
UISourceCode uri to url mappings:
    file:///var/www/html/foo.js -> 
    file:///var/www/bar.js -> 
Adding mapping between network and file system resources.
Emulate reloading inspector.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/bar.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/bar.js.
UISourceCode uri to url mappings:
    file:///var/www/html/foo.js -> http://localhost/html/foo.js
    file:///var/www/bar.js -> http://localhost/bar.js
Removing mapping between network and file system resources.
Emulate reloading inspector.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/bar.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/bar.js.
UISourceCode uri to url mappings:
    file:///var/www/html/foo.js -> 
    file:///var/www/bar.js -> 

Running: testRemoveProject
Adding file system.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/foo.js.
Dumping uiSourceCodes origin URLs:
  - file:///var/www/foo.js
Removing project:
Received project removed event.

Running: testProjectBasedMapping
Adding file system.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/html/foo.js.
Workspace event: Symbol(UISourceCodeAdded): file:///var/www/html2/bar.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/h1/foo.js.
Workspace event: Symbol(UISourceCodeAdded): http://localhost/h2/bar.js.
UISourceCode uri to url mappings:
    file:///var/www/html/foo.js -> http://localhost/h1/foo.js
    file:///var/www/html2/bar.js -> http://localhost/h2/bar.js
